refactor : Home dto res 파일 위치 변경
