+++ src/main/java/com/umc/finly/domain/home/service/HomeWeeklyMoodService.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.home.service;
 
-import com.umc.finly.domain.home.dto.HomeWeeklyMoodRes;
+import com.umc.finly.domain.home.dto.res.HomeWeeklyMoodRes;
 
 public interface HomeWeeklyMoodService {
